requirements.txt
# Core dependency
pillow

# Optional dependencies
# tqdm provides a progress bar in the CLI
# streamlit provides the web GUI
tqdm
streamlit
streamlit-webrtc
numpy
